socketnotifierzmqsocket: Dropped support for REQ-REP protocol in 'SocketNotifierZMQSocket' implementation.
Until now, I dont't have a 'QSocketNotifier' based solution that works with REQ-REP protocol in all conditions. So I decided to drop support until a gerneral solution is available. Correspondingly, I closed gh-2 and created a new feature request gh-4.

Note that 'PollingZMQSocket' supports all kinds of protocols, so if REQ-REP is needed that class can be used instead, which is the default implementation anyway.