Changed fileDescriptor Type to support win_64
Changed Type from qint32 to qintptr, which according to QT "Typedef for
either qint32 or qint64. This type is guaranteed to be the same size as
a pointer on all platforms supported by Qt. On a system with 32-bit
pointers, qintptr is a typedef for qint32; on a system with 64-bit
pointers, qintptr is a typedef for qint64." Solution Taken from
http://stackoverflow.com/questions/18607915/zmq-getsockopt-returns-einval-on-windows-x64-when-local-address-of-zmq-fd-option