FOS-1278: Removes QSocketNotifier for the write end of the ZMQ Socket (#4)
It makes no sense to have a QSocketNotifier for the write end of the ZMQ
socket, especially because most sockets are available for writing most of the
time. With no way to disable this notifier from the public interface for the
socket, the notifier constantly emits the signal that it's available for
writing and it results in it stealing 100% of a single CPU's cycles.